+++ core/domain/src/main/java/com/teampatch/core/domain/usecase/memory/AddMemoryCardUseCase.kt
@@ -1,4 +1,4 @@
-package com.teampatch.core.domain.usecase.home
+package com.teampatch.core.domain.usecase.memory
 
 import com.teampatch.core.domain.model.Image
 import java.time.LocalDateTime
